+++ backend/src/main/java/com/tracepcap/intelligence/dto/WebServerDetailResponse.java
@@ -43,8 +43,8 @@ public class WebServerDetailResponse {
     private int serverErrorCount;
     private String contentType;
     /** frame.number of the first request/response packet (for "view packet"); null if unknown. */
-    private Integer requestFrame;
-    private Integer responseFrame;
+    private Long requestFrame;
+    private Long responseFrame;
   }
 
   /** TLS metadata observed for the server, reconstructed from conversation enrichment. */
